Allow inlining method calls if current inlining is simple (#99)
Bring back inlining the prefix of method calls *iff* the levels inlined so far have all been simple.

By simple, in practice we mean a single method call (rather than chained method calls), or a single method parameter.  This is defined as a property on the Level `OpenOp`s created in `JavaInputAstVisitor`.